separated Drawing, DrawingStateMachine, and Schematic
